Update all_plot function WIP
